fix: build SDK before Next.js on Vercel; add migrations 0024-0025
- Add prebuild script to apps/web that runs sdk:build via --prefix ../..
  so the SDK dist is always generated first, regardless of whether Vercel
  invokes the root build script or apps/web's build script directly
- Migration 0024: Blocks 4A/4B/5/6 — error_fingerprint, v_error_groups,
  v_reliability_kpis, v_user_app_profile, v_user_profile_summary,
  analytics_users columns, and helper functions
- Migration 0025: rename 0023_fix_identity_pipeline to avoid duplicate
  version key; fix GRANT to specify explicit function signature
- Migration 0018: fix ROUND(percentile_cont(...)) by casting to ::numeric

+++ supabase/migrations/0018_performance.sql
@@ -31,11 +31,11 @@ eng as (
 )
 select
   (select count(*) from pl)                                                          as samples,
-  round((select percentile_cont(0.50) within group (order by load_ms) from pl), 0)   as load_p50_ms,
-  round((select percentile_cont(0.75) within group (order by load_ms) from pl), 0)   as load_p75_ms,
-  round((select percentile_cont(0.95) within group (order by load_ms) from pl), 0)   as load_p95_ms,
-  round((select percentile_cont(0.50) within group (order by ttfb_ms) from pl), 0)   as ttfb_p50_ms,
-  round((select percentile_cont(0.50) within group (order by dom_interactive_ms) from pl), 0) as dom_interactive_p50_ms,
+  round((select percentile_cont(0.50) within group (order by load_ms) from pl)::numeric, 0)   as load_p50_ms,
+  round((select percentile_cont(0.75) within group (order by load_ms) from pl)::numeric, 0)   as load_p75_ms,
+  round((select percentile_cont(0.95) within group (order by load_ms) from pl)::numeric, 0)   as load_p95_ms,
+  round((select percentile_cont(0.50) within group (order by ttfb_ms) from pl)::numeric, 0)   as ttfb_p50_ms,
+  round((select percentile_cont(0.50) within group (order by dom_interactive_ms) from pl)::numeric, 0) as dom_interactive_p50_ms,
   round((select avg(engaged_sec) from eng), 0)                                       as avg_engaged_sec;
 
 -- ─── v_performance_by_route ───────────────────────────────────────────────────
@@ -44,8 +44,8 @@ create or replace view public.v_performance_by_route as
 select
   coalesce(nullif(metadata->>'route', ''), '(unknown)')                       as route,
   count(*)                                                                    as samples,
-  round(percentile_cont(0.50) within group (order by (metadata->>'loadMs')::numeric), 0) as load_p50_ms,
-  round(percentile_cont(0.95) within group (order by (metadata->>'loadMs')::numeric), 0) as load_p95_ms,
+  round(percentile_cont(0.50) within group (order by (metadata->>'loadMs')::numeric)::numeric, 0) as load_p50_ms,
+  round(percentile_cont(0.95) within group (order by (metadata->>'loadMs')::numeric)::numeric, 0) as load_p95_ms,
   round(avg((metadata->>'ttfbMs')::numeric), 0)                               as ttfb_avg_ms
 from public.analytics_events
 where name = 'performance.page_load'
@@ -61,7 +61,7 @@ create or replace view public.v_performance_trend as
 select
   date_trunc('day', occurred_at)::date                                        as day,
   count(*)                                                                    as samples,
-  round(percentile_cont(0.50) within group (order by (metadata->>'loadMs')::numeric), 0) as load_p50_ms
+  round(percentile_cont(0.50) within group (order by (metadata->>'loadMs')::numeric)::numeric, 0) as load_p50_ms
 from public.analytics_events
 where name = 'performance.page_load'
   and occurred_at >= now() - interval '14 days'
